% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_table.R
\name{createTable}
\alias{createTable}
\title{create table in SQLite and postgresql databases}
\usage{
createTable(src, table, types, temporary = FALSE, typeUnique = NULL,
  typePK = NULL, typeRequired = NULL, typeDefault = NULL, typeFK = NULL,
  ...)
}
\arguments{
\item{src}{It is the connection as done by \code{dplyr} \code{src_*} group of
methods.}

\item{table}{Name if table. The function will raise error if table name
already exists.}

\item{types}{Named vector of column names and type. For \code{character} data
type, we can add \code{"character(30)"} to denote length of column. Common
types are \code{"integer", "numeric", "character", "date"}. We can also add
\code{"TIMESTAMP"} and \code{"SERIAL"} data type in postgresql for
\code{timestamp, with default CURRENT_TIMESTAMP} and \code{primary key with
autonumbering}.}

\item{temporary}{(default \code{FALSE}). It denotes whether the table is a
temporary one.}

\item{typeUnique}{character vector denoting the columns with only unique
values allowed.}

\item{typePK}{character vector (usually, single string) denoting the Primary
Key of the table.}

\item{typeRequired}{character vector denoting the columns which cannot be
left empty.}

\item{typeDefault}{named list with list(column1 = "defaultVal1", column2 =
"defaultval2", ...) format.}

\item{typeFK}{named list of the format list(fk_column1 = "tableName1withPK",
fk_column2 = "tableName2withPK", ...)}

\item{onUpdate}{to be used only if foreign key is defined. One of the
following options: \code{"cascade", "noAction", "setNull"}.}

\item{onDelete}{to be used only if foreign key is defined. One of the
following options: \code{"cascade", "noAction", "setNull"}.}
}
\value{
Resulting SQL statement.
}
\description{
create table in SQLite and postgresql databases
}
\examples{
\dontrun{
createTable(src, table = "tbl1", types = c(id = "numeric", name =
"character(50)", address = "character(50)", age = "numeric"), temporary = F,
typeUnique = c("name", "address"), typePK = "id", typeRequired = "name",
typeDefault = list(age = 20, name = "mee"))

createTable(src, table = "tbl2", types = c(id = "numeric", fk_id  =
"numeric", name = "character(50)", address = "character(50)", age =
"numeric"), temporary = F, typeUnique = "name", typePK = "id", typeRequired =
c("fk_id", "name"), typeDefault = list(age = 20), typeFK = list(fk_id =
"tbl1"), onUpdate = "setNull", onDelete = "cascade")
}
}

