% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPK.R
\name{getPK}
\alias{getPK}
\title{Get Primary Key of a table and find the next available value for the same.}
\usage{
getPK(src, table)
}
\arguments{
\item{src}{The \code{src_*} functions for database (postgresql and sqlite) from \code{dplyr} package.}

\item{table}{String representing table name of the data base.}
}
\value{
Returns a data frame with table name, column name, type of data and next available number as columns.
}
\description{
Get Primary Key of a table and find the next available value for the same.
}

