% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nextNumber}
\alias{nextNumber}
\title{Get next number for PK generation
In a vector of numbers, the function finds out the first missing number and returns it.
If it does not find any missing value, it return number next to the last number.}
\usage{
nextNumber(x)
}
\arguments{
\item{x}{vector of numbers.}
}
\description{
Get next number for PK generation
In a vector of numbers, the function finds out the first missing number and returns it.
If it does not find any missing value, it return number next to the last number.
}

